Exception handling

+++ src/main/java/com/victoroliveira/dscommerce/services/ProductService.java
@@ -1,16 +1,20 @@
 package com.victoroliveira.dscommerce.services;
 
-import java.util.Optional;
-
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.victoroliveira.dscommerce.dto.ProductDTO;
 import com.victoroliveira.dscommerce.entities.Product;
 import com.victoroliveira.dscommerce.repositories.ProductRepository;
+import com.victoroliveira.dscommerce.services.exceptions.DatabaseException;
+import com.victoroliveira.dscommerce.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -19,12 +23,11 @@ public class ProductService {
 	private ProductRepository repository;
 	
 	@Transactional(readOnly = true)
-	public ProductDTO findById(Long id) {
+	public ProductDTO findById(Long id) {			
 		
-		Optional<Product> result = repository.findById(id);
-		Product product = result.get();		
-		ProductDTO dto = new ProductDTO(product);		
-		return dto;
+		Product product = repository.findById(id).orElseThrow(
+				()-> new ResourceNotFoundException("Recurso não encontrado"));			
+		return new ProductDTO(product);		
 		
 	}
 	
@@ -47,18 +50,28 @@ public class ProductService {
 	
 	@Transactional
 	public ProductDTO update(Long id, ProductDTO dto) {		
-		
+		try {
 		Product entity = repository.getReferenceById(id);		
 		copyDtoToEntity(dto, entity);		
 		entity = repository.save(entity);		
 		return new ProductDTO(entity);
-		
+		}
+		catch(EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
 	}
 	
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void deleteById(Long id) {		
-		repository.deleteById(id);			
-		
+		if(!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+		try {
+		repository.deleteById(id);	
+		}
+		catch(DataIntegrityViolationException e) {
+			throw new DatabaseException("Falha de integridade referencial");
+		}
 	}
 
 	private void copyDtoToEntity(ProductDTO dto, Product entity) {
